refactor(core): reduce duplication in fetchWithRetry and stream-until-idle
- Extracted `fetchWithRetry` to its own file to avoid duplicated implementation in `packages/core/src/utils.ts`.
- Extracted shared functions `resolveStreamUntilIdleScope`, `buildContinuationDirective`, `buildContinuationOpts`, `acquireStreamSlot`, and `releaseStreamSlot` from `packages/core/src/agent/stream-until-idle.ts` and `packages/core/src/agent/durable/durable-stream-until-idle.ts` into a new `packages/core/src/agent/shared/stream-until-idle.ts` helper file.
